+++ fluxcloud/main/client.py
@@ -5,11 +5,10 @@
 
 import os
 import shutil
-import time
 
+import fluxcloud.main.api as api
 import fluxcloud.utils as utils
 from fluxcloud.logger import logger
-from fluxcloud.main.api import APIClient
 from fluxcloud.main.decorator import save_meta, timed
 
 here = os.path.dirname(os.path.abspath(__file__))
@@ -93,80 +92,6 @@ class ExperimentClient:
         raise NotImplementedError
 
     @save_meta
-    def open_ui(self, setup, experiment, size, api=None, persistent=False):
-        """
-        Launch a CRD that opens the UI only.
-        """
-        # The MiniCluster can vary on size
-        minicluster = experiment.minicluster
-
-        # Create a FluxRestful API to submit to
-        created = False
-        if api is None:
-            api = APIClient()
-            created = True
-
-        logger.info(f"\n🌀 Bringing up MiniCluster of size {size}")
-
-        # Get persistent variables for this job size, image is required
-        job = experiment.get_persistent_variables(size, required=["image"])
-        job.update({"token": api.token, "user": api.user})
-
-        # We can't have a command
-        if "command" in job:
-            del job["command"]
-
-        # Pre-pull containers, etc.
-        if hasattr(self, "pre_apply"):
-            self.pre_apply(experiment, "global-job", job=job)
-
-        # Create the minicluster via a CRD without a command
-        crd = experiment.generate_crd(job, size)
-
-        # Create one MiniCluster CRD (without a command) to run the Flux Restful API
-        kwargs = {
-            "minicluster": minicluster,
-            "crd": crd,
-            "token": api.token,
-            "user": api.user,
-            "size": size,
-        }
-        submit_script = experiment.get_shared_script(
-            "minicluster-create-persistent", kwargs, suffix=f"-size-{size}"
-        )
-        # Start the MiniCluster! This should probably be done better...
-        self.run_timed(
-            f"minicluster-create-persistent-size-{size}", ["/bin/bash", submit_script]
-        )
-
-        # Ensure our credentials still work, and open port forward
-        api.check(experiment)
-        logger.info(f"\n🌀 MiniCluster of size {size} is up.\n")
-
-        # If created for the first time, show credentials
-        if created:
-            logger.info(
-                "Save these if you want to log into the Flux RESTFul interface, there are specific to the MiniCluster"
-            )
-            logger.info(f"export FLUX_USER={api.user}")
-            logger.info(f"export FLUX_TOKEN={api.token}")
-
-        # If we exit, the port forward will close.
-        if persistent:
-            try:
-                logger.info("Press Control+c to Disconnect.")
-                while True:
-                    time.sleep(10)
-            except KeyboardInterrupt:
-                logger.info("🧽️ Cleaning up!")
-                self.run_timed(
-                    f"minicluster-persistent-destroy-size-{size}",
-                    ["kubectl", "delete", "-f", crd],
-                )
-
-        return api, kwargs
-
-    @save_meta
     def submit(self, setup, experiment):
         """
         Submit a Job via the Restful API
@@ -177,8 +102,6 @@ class ExperimentClient:
             )
             return
 
-        api = None
-
         # Iterate through all the cluster sizes
         for size in experiment.minicluster["size"]:
             # We can't run if the minicluster > the experiment size
@@ -188,21 +111,47 @@ class ExperimentClient:
                 )
                 continue
 
-            # Open the api for the size
-            api, uiattrs = self.open_ui(setup, experiment, size, api)
-            logger.info(f"\n🌀 Bringing up MiniCluster of size {size}")
+            # Launch a unique Minicluster per container image. E.g.,
+            # if the user provides 2 images for size 4, we create two MiniClusters
+            for minicluster in experiment.get_submit_miniclusters(size):
+                logger.info(
+                    f"\n🌀 Bringing up MiniCluster of size {size} with image {minicluster['image']}"
+                )
 
-            # Save times (and logs in submit) as we go
-            for jobid, info in api.submit(setup, experiment, size):
-                logger.info(f"{jobid} took {info['runtime']} seconds.")
-                self.times[jobid] = info["runtime"]
-                self.info[jobid] = info
+                # Create the API client (creates the user and token for the cluster)
+                cli = api.APIClient()
 
-            logger.info(f"\n🌀 MiniCluster of size {size} is finished")
-            self.run_timed(
-                f"minicluster-persistent-destroy-size-{size}",
-                ["kubectl", "delete", "-f", uiattrs["crd"]],
-            )
+                # Pre-pull containers, etc.
+                if hasattr(self, "pre_apply"):
+                    self.pre_apply(
+                        experiment,
+                        minicluster["name"],
+                        job={"image": minicluster["image"]},
+                    )
+
+                # Get back results with times (for minicluster assets) and jobs
+                results = cli.submit(setup, experiment, minicluster)
+
+                # Save times and output files for jobs
+                for job in results.get("jobs", []):
+                    self.save_job(job)
+
+    def save_job(self, job):
+        """
+        Save the job and add times to our times listing.
+        """
+        jobid = f"{job['id']}"
+        self.times[jobid] = job["info"]["runtime"]
+
+        # Do we have an output file and output?
+        if job["output"]:
+            # Save to our output directory!
+            logfile = job["job_output"]
+            utils.mkdir_p(os.path.dirname(logfile))
+            utils.write_file(job["output"], logfile)
+
+        del job["output"]
+        self.info[jobid] = job
 
     @save_meta
     def apply(self, setup, experiment):